lib: build a state machine for control transfers
The lm4f120 hardware handles the status stage rather than giving me the
raw packets, so I have to keep track of the state machine to figure out
what packets the hardware received and recreate those packets (i.e. the
zero-length status OUT) for the usb_device interface, and vice-versa
(i.e. turning the zero-length status IN into setting the DATAEND bit).

Yes, this means we have

                    USB packets
                         |
                         v
    the hardware's control transfer state machine
                         |
                         v
                 struct Endpoint0
     which reverse-engineers the previous and turns
              them back into USB packets
                         |
                         v
              usb_device::control_pipe
        which turns those packets back into a
           control transfer state machine

and if those state machines are either (a) not invertible, or (b) not
exactly matched ... this is going to be full of bugs.  But I think it's
the best I can do to match this hardware to this API.